refactor: env.ts를 index.ts로 변경
